Start to refactor the IOPs into something more generally useful
