+++ include/felspar/poll/warden.hpp
@@ -1,34 +1,11 @@
 #pragma once
 
+#include <felspar/poll/completion.hpp>
 
-#include <felspar/coro/task.hpp>
-
-#include <functional>
-#include <map>
 #include <vector>
-
-#include <poll.h>
 
 
 namespace felspar::poll {
-
-
-    template<typename R>
-    struct iop;
-    class warden;
-
-
-    template<>
-    struct iop<void> {
-        /// TODO Want something with much lower overhead than std::function
-        std::function<void(felspar::coro::coroutine_handle<>)> suspend;
-
-        bool await_ready() const noexcept { return false; }
-        void await_suspend(felspar::coro::coroutine_handle<> h) noexcept {
-            suspend(h);
-        }
-        auto await_resume() noexcept {}
-    };
 
 
     /// Executor that allows `poll` based asynchronous IO
@@ -58,6 +35,14 @@ namespace felspar::poll {
             live.push_back(std::move(coro));
         }
 
+        /**
+         * ### Socket APIs
+         */
+        //         virtual iop<int> accept(int fd) = 0;
+
+        /**
+         * ### File readiness
+         */
         virtual iop<void> read_ready(int fd) = 0;
         virtual iop<void> write_ready(int fd) = 0;
     };
